New Skills, Wrapper Functionality
Added 2H Piercing, Remove Traps and Tripple Attack to the skills page.

There were stylesheet naming collisions when you would wrap charbrowser
in your forum header/footer. For example, we used the "*" selector to
set all margins and padding to 0px. That would potentially distort your
forums margins. To correct it the entire body of the charbrowser has
been wrapped in a div with an id of "charbrowser" and a #charbrowser
prefix has been added to every css selector. This has broad implications
but seems to have worked well; please let me know if you've noticed any
changes with the display style.