v2.6.6 corrected general slot order
The bags were numbered top to bottom then left to right. They should
have been left to right then top to bottom. This has apparently always
been wrong but was noticed while adding uleats inventory changes.

Fixed the missing 'px' on the bagbuttonrow styles. This was causing the
"Done" button to float to the top of the bag window in some uncommon
scenarios (when wrapped).